Checkpoint: implementing recursive host test
Make MockHost a lot smarter. Add a program map.
Refactor some test-related and interface code, consolidate things in
interface crate.
Add AthenaRevision.
Allow MockHost to recursively call into the VM. Required some memory
magic.
