+++ runner/src/host.rs
@@ -30,114 +30,3 @@ impl From<u64> for Bytes32AsU64 {
     Bytes32AsU64(bytes)
   }
 }
-
-// currently unused
-#[derive(Debug, Clone, Copy)]
-pub enum AthenaCapability {}
-
-// currently unused
-#[derive(Debug, Clone)]
-pub enum AthenaOption {}
-
-#[derive(Debug)]
-pub enum SetOptionError {
-  InvalidKey,
-  InvalidValue,
-}
-
-#[cfg(test)]
-use std::collections::BTreeMap;
-
-#[cfg(test)]
-use athena_interface::{
-  Address, AthenaMessage, ExecutionResult, HostInterface, StorageStatus, TransactionContext,
-};
-
-#[cfg(test)]
-pub(crate) struct MockHost {
-  context: Option<TransactionContext>,
-
-  // stores state keyed by address and key
-  storage: BTreeMap<(Address, Bytes32), Bytes32>,
-
-  // stores balance keyed by address
-  balance: BTreeMap<Address, Bytes32>,
-}
-
-#[cfg(test)]
-impl MockHost {
-  pub fn new(context: Option<TransactionContext>) -> Self {
-    MockHost {
-      context: context,
-      storage: BTreeMap::new(),
-      balance: BTreeMap::new(),
-    }
-  }
-}
-
-#[cfg(test)]
-impl HostInterface for MockHost {
-  fn account_exists(&self, addr: &Address) -> bool {
-    self.balance.contains_key(addr)
-  }
-
-  // return null bytes if the account or key do not exist
-  fn get_storage(&self, address: &Address, key: &Bytes32) -> Bytes32 {
-    *self
-      .storage
-      .get(&(*address, *key))
-      .unwrap_or(&Bytes32::default())
-  }
-
-  fn set_storage(&mut self, address: &Address, key: &Bytes32, value: &Bytes32) -> StorageStatus {
-    self.storage.insert((*address, *key), *value);
-    return StorageStatus::StorageAssigned;
-  }
-
-  // return 0 if the account does not exist
-  fn get_balance(&self, address: &Address) -> u64 {
-    self
-      .balance
-      .get(address)
-      .map_or_else(|| 0, |balance| Bytes32AsU64(*balance).into())
-  }
-
-  fn get_tx_context(&self) -> TransactionContext {
-    self.context.unwrap()
-  }
-
-  fn get_block_hash(&self, _number: i64) -> Bytes32 {
-    Bytes32::default()
-  }
-
-  fn call(&mut self, _msg: AthenaMessage) -> ExecutionResult {
-    ExecutionResult::new(athena_interface::StatusCode::Failure, 0, None, None)
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use super::*;
-
-  #[test]
-  fn test_get_storage() {
-    let mut host = MockHost::new(None);
-    let address = [8; 24];
-    let key = [1; 32];
-    let value = [2; 32];
-    assert_eq!(
-      host.set_storage(&address, &key, &value),
-      StorageStatus::StorageAssigned
-    );
-    let retrieved_value = host.get_storage(&address, &key);
-    assert_eq!(retrieved_value, value);
-  }
-
-  #[test]
-  fn test_get_balance() {
-    let host = MockHost::new(None);
-    let address = [8; 24];
-    let balance = host.get_balance(&address);
-    assert_eq!(balance, 0);
-  }
-}
