+++ runner/src/lib.rs
@@ -2,4 +2,4 @@ pub mod host;
 pub mod vm;
 
 pub use host::Bytes32AsU64;
-pub use vm::{AthenaVm, VmInterface};
+pub use vm::AthenaVm;
