+++ tests/recursive_call/src/main.rs
@@ -1,10 +1,12 @@
 #![no_main]
 athena_vm::entrypoint!(main);
 
-use athena_vm::helpers::{address_to_32bit_words, address_to_bytes32, bytes32_to_32bit_words};
+use athena_vm::helpers::{address_to_32bit_words, bytes32_to_32bit_words};
 use athena_vm::types::{
   StorageStatus::StorageAdded, StorageStatus::StorageModified, ADDRESS_ALICE, HELLO_WORLD,
 };
+
+// Note: the test harness installs this contract code at ADDRESS_ALICE
 
 // storage status is returned as a 256-bit value
 const STORAGE_ADDED: [u32; 8] = [StorageAdded as u32, 0, 0, 0, 0, 0, 0, 0];
@@ -17,7 +19,10 @@ fn return_value(value: u32) {
   let val: [u32; 8] = [value, 0, 0, 0, 0, 0, 0, 0];
   let address = address_to_32bit_words(ADDRESS_ALICE);
   unsafe { athena_vm::host::write_storage(key.as_mut_ptr(), address.as_ptr(), val.as_ptr()) };
-  assert_eq!(key, STORAGE_MODIFIED, "write_storage failed");
+  assert!(
+    key == STORAGE_ADDED || key == STORAGE_MODIFIED,
+    "write_storage failed"
+  );
 }
 
 fn recursive_call(value: u32) -> u32 {
